azure-function/requirements.txt
# Azure Functions for Python
azure-functions
requests
